Reorganize the strucutre of source files in execution engine.
