--- execution-engine/src/hcall/wasi/mod.rs
@@ -1,14 +1,0 @@
-//! An implementation of the WASI API for Execution Engine.
-//!
-//! ## Authors
-//!
-//! The Veracruz Development Team.
-//!
-//! ## Licensing and copyright notice
-//!
-//! See the `LICENSE.markdown` file in the Veracruz root directory for
-//! information on licensing and copyright.
-
-pub mod wasmi;
-#[cfg(feature = "std")]
-pub mod wasmtime;
