--- execution-engine/src/hcall/wasi/wasmi/mod.rs
@@ -1,12 +1,0 @@
-//! An implementation of the WebAssembly System Interface for WASMI.
-//!
-//! # Authors
-//!
-//! The Veracruz Development Team.
-//!
-//! # Copyright
-//!
-//! See the file `LICENSE.markdown` in the Veracruz root directory for licensing
-//! and copyright information.
-
-pub mod runtime_state;
